backend/requirements.txt
asgiref==3.11.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
Django==6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.45.0
h11==0.16.0
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.2
uvicorn==0.40.0
